[Feature] #81 - 게시글 상세 페이지를 컴포넌트로 수정
